buffer-overflow/requirements.txt
colorama==0.4.6
telnetlib3==2.0.4
